Singular costmap publisher for simulation
