+++ urc_test/setup.cfg
@@ -1,4 +1,4 @@
 [develop]
-script_dir=$base/lib/tester
+script_dir=$base/lib/urc_test
 [install]
-install_scripts=$base/lib/tester
+install_scripts=$base/lib/urc_test
